+++ Cargo.toml
@@ -1,11 +1,28 @@
 [package]
 name = "gravity"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
+rust-version = "1.72"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 rand = "0.8.5"
-eframe = "0.24.1"
-rayon = "1.5"
+# enable the js feature in getrandom
+getrandom = { features = ["js"] }
+eframe = {version = "0.24.1", default-features = false, features = [
+    "default_fonts", # Embed the default egui fonts.
+    "glow",          # Use the glow rendering backend. Alternative: "wgpu".
+]}
+rayon = "1.8"
+
+# native:
+#[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+#env_logger = "0.10"
+
+# web:
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen-futures = "0.4"
+
+[profile.release]
+opt-level = 2 # fast and small wasm
