+++ assets/manifest.json
@@ -1,0 +1,10 @@
+{
+  "name": "gravity",
+  "short_name": "vidstge-gravity",
+  "lang": "en-US",
+  "id": "/index.html",
+  "start_url": "./index.html",
+  "display": "standalone",
+  "background_color": "white",
+  "theme_color": "white"
+}
